+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Notify user on CCv2 settings validation errors [#1068](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1068)
 - Migrated `SAP CCM` logic to Kotlin coroutines [#1069](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1069)
 - Introduced `Native` integration via OpenAPI [#1070](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1070)
+- Added possibility to change Integration Protocol and current subscription [#1071](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1071)
 
 ### `Type System` enhancements
 - Show Item line marker for jalo classes [#1067](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1067)